requirements.txt
numpy==1.21.6
opencv_python==4.5.5.64
Pillow==9.2.0
scikit_learn==1.1.1
